updae

+++ basetype.h
@@ -8,16 +8,15 @@
 
 #ifndef BaseOpenGL_basetype_h
 #define BaseOpenGL_basetype_h
-
-//======= 定义数据类型 ========
-typedef long			int64;
-typedef unsigned long	uint64;
-typedef int				int32;
-typedef unsigned int	uint32;
-typedef short			int16;
-typedef unsigned short	uint16;
-typedef char			int8;
-typedef unsigned char	uint8;
-//=============================
+//========= 定义数据类型 ==========
+typedef long long			int64;
+typedef unsigned long long	uint64;
+typedef int					int32;
+typedef unsigned int		uint32;
+typedef short				int16;
+typedef unsigned short		uint16;
+typedef char				int8;
+typedef unsigned char		uint8;
+//=================================
 
 #endif
